Move the "loadAndCheck"-code into the new "LoadAndCheckUrls"-class.
